Dash Updated & Renamed States
Added a goal tracker, daily calorie counter, and added buttons for each
food in the database that adds the calorie count for a single serving to
the daily calorie counter.

Next: Add calories spent functionality and build an array of all the
meals eaten for the day on the dash.